+++ src/glsl_shader.h
@@ -35,7 +35,7 @@ public:
   void UseProgram() override;
 
   void SetInt(const std::string& name, int value) override;
-  void SetFloat(const std::string& name, int value) override;
+  void SetFloat(const std::string& name, float value) override;
   void SetVec3(const std::string& name, const glm::vec3& value) override;
   void SetVec4(const std::string& name, const glm::vec4& value) override;
   void SetMat4(const std::string& name, const glm::mat4& value) override;
